+++ application/src/main/java/org/ehrbase/application/EhrBase.java
@@ -20,7 +20,7 @@ package org.ehrbase.application;
 
 import org.ehrbase.ServiceModuleConfiguration;
 import org.ehrbase.rest.ehrscape.RestEHRScapeModuleConfiguration;
-import org.ehrbase.rest.openehr.RestOpenehrModuleConfiguration;
+import org.ehrbase.rest.RestModuleConfiguration;
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
 import org.springframework.context.annotation.Import;
@@ -28,7 +28,7 @@ import org.springframework.context.annotation.Import;
 @SpringBootApplication()
 @Import({ServiceModuleConfiguration.class,
         RestEHRScapeModuleConfiguration.class,
-        RestOpenehrModuleConfiguration.class,
+        RestModuleConfiguration.class,
 })
 public class EhrBase {