+++ rest-openehr/src/main/java/org/ehrbase/rest/admin/AdminController.java
@@ -16,14 +16,14 @@
  * limitations under the License.
  */
 
-package org.ehrbase.rest.openehr.controller.admin;
+package org.ehrbase.rest.admin;
 
 import io.swagger.annotations.Api;
 import io.swagger.annotations.ApiResponse;
 import io.swagger.annotations.ApiResponses;
 import io.swagger.annotations.ResponseHeader;
 import org.ehrbase.response.openehr.admin.AdminStatusResponseData;
-import org.ehrbase.rest.openehr.controller.BaseController;
+import org.ehrbase.rest.BaseController;
 import org.springframework.boot.autoconfigure.condition.ConditionalOnProperty;
 import org.springframework.http.MediaType;
 import org.springframework.http.ResponseEntity;
@@ -34,8 +34,8 @@ import org.springframework.web.bind.annotation.RestController;
 @Api(tags = {"Admin", "Heartbeat"})
 @ConditionalOnProperty(prefix = "admin-api", name = "active")
 @RestController
-@RequestMapping(path = "/rest/openehr/v1/admin", produces = {MediaType.APPLICATION_JSON_VALUE, MediaType.APPLICATION_XML_VALUE})
-public class OpenehrAdminController extends BaseController {
+@RequestMapping(path = "/rest/admin", produces = {MediaType.APPLICATION_JSON_VALUE, MediaType.APPLICATION_XML_VALUE})
+public class AdminController extends BaseController {
 
     @GetMapping(path = "/status")
     @ApiResponses(value = {