+++ CHANGELOG.md
@@ -13,12 +13,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Support for Attribute-based Access Control (see: https://github.com/ehrbase/ehrbase/pull/499)
 
 ### Changed
-
+- Update paths for Admin API, Management API and `/status` endpoint (see: https://github.com/ehrbase/ehrbase/pull/541) 
 
 ### Fixed
 - Folder handling (update, delete and missing audits) (see: https://github.com/ehrbase/ehrbase/pull/529)
-
-
 
 ## [0.16.0] (beta)